Closes #3357: split verify_execution.rs into a module directory
Pure mechanical reorg of the 2533-LOC crates/henyey/src/verify_execution.rs
into a verify_execution/ directory, with no logic change:

- mod.rs: the 3 shared structs (VerifyExecutionOptions / VerifyContext /
  VerifyStats), cmd_verify_execution, and setup (kept here, next to the
  VerifyContext it constructs, to avoid struct-literal field-visibility churn)
- diffs.rs: the 11 diagnostic/diff printers, marked pub(super)
- run.rs: the verification loop, the #3339 pure helpers, verify_single_ledger
  (with the hash_mismatch_debug tracing::warn! blocks verbatim), summary/
  eviction diagnostics, and the entire #[cfg(test)] mod tests

Visibility deltas are minimal: pub(super) on the 11 printers + the two run
entry points (run_verification_loop, print_summary) called from mod.rs. No
struct-field visibility bumps were needed — run is a descendant module of
verify_execution, so it accesses the parent structs' private fields directly.

verify-execution is the Henyey-only offline hash-verification harness
(PARITY_STATUS.md:166-168, does not affect parity %). The move is byte-stable:
all println!/bail!/WARN sites and the hash_mismatch_debug WARN target/fields/
message are unchanged, so verify output, VerifyStats counters, and print_summary
text are identical. The 10 verify_execution tests keep passing under
cargo test -p henyey with an identical test-name set (now under
verify_execution::run::tests).

Refs #3357
