+++ crates/henyey/src/verify_execution/run.rs
@@ -1,1088 +1,28 @@
-//! Offline execution verification — replays ledgers against CDP data.
+//! Phase 5/6: the per-ledger verification loop, its pure helpers, the
+//! `verify_single_ledger` orchestrator, and the run/eviction diagnostics.
+//!
+//! Moved verbatim from the former single-file module (#3357 split); no logic
+//! change. The `hash_mismatch_debug` WARN blocks are byte-identical to the
+//! pre-split source.
 
-use std::sync::Arc;
 use std::time::{Duration, Instant};
 
 use stellar_xdr::curr::WriteXdr;
 
-use henyey_app::AppConfig;
-use henyey_bucket::{BucketList, BucketManager, HotArchiveBucketList, PendingMergeState};
 use henyey_common::Hash256;
 use henyey_history::cdp::{
     extract_ledger_close_data, extract_ledger_header, extract_transaction_results,
-    CachedCdpDataLake,
 };
 use henyey_history::checkpoint;
 use henyey_history::verify;
-use henyey_ledger::{LedgerManager, LedgerManagerConfig};
 
-pub(crate) struct VerifyExecutionOptions {
-    pub from: Option<u32>,
-    pub to: Option<u32>,
-    pub stop_on_error: bool,
-    pub show_diff: bool,
-    pub cdp_url: Option<String>,
-    pub cdp_date: Option<String>,
-    pub cache_dir: Option<std::path::PathBuf>,
-    pub no_cache: bool,
-    pub quiet: bool,
-}
-
-/// Bundles all long-lived resources needed across verification phases.
-pub(crate) struct VerifyContext {
-    archive: henyey_history::HistoryArchive,
-    cdp: CachedCdpDataLake,
-    ledger_manager: LedgerManager,
-    _bucket_manager: Arc<BucketManager>,
-    // TempDir guards — dropping these would delete the temp directories.
-    _cdp_dir_holder: Option<tempfile::TempDir>,
-    _bucket_dir_holder: Option<tempfile::TempDir>,
-    // Configuration
-    start_ledger: u32,
-    end_ledger: u32,
-    init_checkpoint: u32,
-    end_checkpoint: u32,
-    init_header_hash: Hash256,
-    stop_on_error: bool,
-    show_diff: bool,
-    quiet: bool,
-}
-
-/// Accumulator counters for the verification run.
-#[derive(Default)]
-pub(crate) struct VerifyStats {
-    pub ledgers_verified: u32,
-    pub ledgers_matched: u32,
-    pub ledgers_mismatched: u32,
-    pub header_mismatches: u32,
-    pub tx_result_mismatches: u32,
-    pub meta_mismatches: u32,
-    total_close_us: u64,
-    total_tx_exec_us: u64,
-    total_commit_us: u64,
-    total_add_batch_us: u64,
-    total_eviction_us: u64,
-    total_tx_count: usize,
-    total_cache_hits: u64,
-    total_cache_misses: u64,
-    slowest_ledger_us: u64,
-    slowest_ledger_seq: u32,
-    slowest_txs: Vec<(u32, String, u64)>,
-    peak_rss_bytes: u64,
-}
-
-/// Returns a short description of a TransactionEnvelope for diagnostics:
-/// (op_types_csv, declared_fee, soroban_resource_fee, soroban_resources_summary).
-fn describe_envelope(
-    env: Option<&stellar_xdr::curr::TransactionEnvelope>,
-) -> (String, i64, i64, String) {
-    use stellar_xdr::curr::{
-        Operation, OperationBody, SorobanTransactionData, TransactionEnvelope, TransactionExt,
-    };
-    fn describe_op(op: &Operation) -> &'static str {
-        match op.body {
-            OperationBody::CreateAccount(_) => "CreateAccount",
-            OperationBody::Payment(_) => "Payment",
-            OperationBody::PathPaymentStrictReceive(_) => "PathPayStrictRecv",
-            OperationBody::ManageSellOffer(_) => "ManageSellOffer",
-            OperationBody::CreatePassiveSellOffer(_) => "CreatePassiveSellOffer",
-            OperationBody::SetOptions(_) => "SetOptions",
-            OperationBody::ChangeTrust(_) => "ChangeTrust",
-            OperationBody::AllowTrust(_) => "AllowTrust",
-            OperationBody::AccountMerge(_) => "AccountMerge",
-            OperationBody::Inflation => "Inflation",
-            OperationBody::ManageData(_) => "ManageData",
-            OperationBody::BumpSequence(_) => "BumpSequence",
-            OperationBody::ManageBuyOffer(_) => "ManageBuyOffer",
-            OperationBody::PathPaymentStrictSend(_) => "PathPayStrictSend",
-            OperationBody::CreateClaimableBalance(_) => "CreateClaimableBalance",
-            OperationBody::ClaimClaimableBalance(_) => "ClaimClaimableBalance",
-            OperationBody::BeginSponsoringFutureReserves(_) => "BeginSponsoringFutureReserves",
-            OperationBody::EndSponsoringFutureReserves => "EndSponsoringFutureReserves",
-            OperationBody::RevokeSponsorship(_) => "RevokeSponsorship",
-            OperationBody::Clawback(_) => "Clawback",
-            OperationBody::ClawbackClaimableBalance(_) => "ClawbackClaimableBalance",
-            OperationBody::SetTrustLineFlags(_) => "SetTrustLineFlags",
-            OperationBody::LiquidityPoolDeposit(_) => "LiquidityPoolDeposit",
-            OperationBody::LiquidityPoolWithdraw(_) => "LiquidityPoolWithdraw",
-            OperationBody::InvokeHostFunction(_) => "InvokeHostFunction",
-            OperationBody::ExtendFootprintTtl(_) => "ExtendFootprintTtl",
-            OperationBody::RestoreFootprint(_) => "RestoreFootprint",
-        }
-    }
-    fn summarize_resources(data: &SorobanTransactionData) -> String {
-        use stellar_xdr::curr::SorobanTransactionDataExt;
-        let r = &data.resources;
-        let archived = match &data.ext {
-            SorobanTransactionDataExt::V1(ext) => ext.archived_soroban_entries.len(),
-            _ => 0,
-        };
-        // archived_idx: count of read-write footprint entries marked for
-        // hot-archive autorestore (see CAP-66/77). Surfaced for diagnostic
-        // parity with stellar-core's restore-then-meter pre-execution path.
-        format!(
-            "instr={},rro={},rrw={},wb={},drb={},archived_idx={}",
-            r.instructions,
-            r.footprint.read_only.len(),
-            r.footprint.read_write.len(),
-            r.write_bytes,
-            r.disk_read_bytes,
-            archived,
-        )
-    }
-    let Some(env) = env else {
-        return (String::from("?"), 0, 0, String::new());
-    };
-    match env {
-        TransactionEnvelope::TxV0(v0) => {
-            let ops: Vec<&str> = v0.tx.operations.iter().map(describe_op).collect();
-            (ops.join(","), v0.tx.fee as i64, 0, String::new())
-        }
-        TransactionEnvelope::Tx(v1) => {
-            let ops: Vec<&str> = v1.tx.operations.iter().map(describe_op).collect();
-            let (sb_fee, sb_summary) = match &v1.tx.ext {
-                TransactionExt::V1(data) => (data.resource_fee, summarize_resources(data)),
-                _ => (0, String::new()),
-            };
-            (ops.join(","), v1.tx.fee as i64, sb_fee, sb_summary)
-        }
-        TransactionEnvelope::TxFeeBump(fb) => {
-            let (inner_ops_str, inner_sb_fee, inner_sb_summary) = match &fb.tx.inner_tx {
-                stellar_xdr::curr::FeeBumpTransactionInnerTx::Tx(inner) => {
-                    let ops_str = inner
-                        .tx
-                        .operations
-                        .iter()
-                        .map(describe_op)
-                        .collect::<Vec<_>>()
-                        .join(",");
-                    let (sb_fee, sb_summary) = match &inner.tx.ext {
-                        TransactionExt::V1(data) => (data.resource_fee, summarize_resources(data)),
-                        _ => (0, String::new()),
-                    };
-                    (ops_str, sb_fee, sb_summary)
-                }
-            };
-            (
-                format!("FB({inner_ops_str})"),
-                fb.tx.fee,
-                inner_sb_fee,
-                inner_sb_summary,
-            )
-        }
-    }
-}
-
-/// Extract a CSV summary of soroban diagnostic events from a transaction's
-/// meta. Captures the SCEC_EXCEEDED_LIMIT errors that pinpoint which
-/// stellar-core resource check fired (e.g.
-/// "operation byte-write resources exceeds amount specified") — critical
-/// for #2503 hash-mismatch divergence diagnosis.
-fn extract_diagnostic_event_summary(meta: Option<&stellar_xdr::curr::TransactionMeta>) -> String {
-    use stellar_xdr::curr::{ScVal, TransactionMeta};
-    let Some(meta) = meta else {
-        return String::new();
-    };
-    let events: &[stellar_xdr::curr::DiagnosticEvent] = match meta {
-        TransactionMeta::V3(m) => m
-            .soroban_meta
-            .as_ref()
-            .map(|s| s.diagnostic_events.as_slice())
-            .unwrap_or(&[]),
-        TransactionMeta::V4(m) => m.diagnostic_events.as_slice(),
-        _ => return String::new(),
-    };
-    let mut summaries = Vec::new();
-    for event in events {
-        let body = match &event.event.body {
-            stellar_xdr::curr::ContractEventBody::V0(b) => b,
-        };
-        let mut parts: Vec<String> = Vec::new();
-        for topic in body.topics.as_slice() {
-            match topic {
-                ScVal::Symbol(s) => parts.push(s.to_utf8_string_lossy()),
-                ScVal::Error(e) => parts.push(format!("{:?}", e)),
-                ScVal::String(s) => parts.push(s.to_utf8_string_lossy()),
-                _ => parts.push(format!("{:?}", std::mem::discriminant(topic))),
-            }
-        }
-        match &body.data {
-            ScVal::Vec(Some(v)) => {
-                let inner: Vec<String> =
-                    v.0.as_slice()
-                        .iter()
-                        .filter_map(|x| match x {
-                            ScVal::U64(n) => Some(n.to_string()),
-                            ScVal::U32(n) => Some(n.to_string()),
-                            ScVal::I64(n) => Some(n.to_string()),
-                            ScVal::I32(n) => Some(n.to_string()),
-                            ScVal::Symbol(s) => Some(s.to_utf8_string_lossy()),
-                            ScVal::String(s) => Some(s.to_utf8_string_lossy()),
-                            _ => None,
-                        })
-                        .collect();
-                if !inner.is_empty() {
-                    parts.push(format!("[{}]", inner.join(",")));
-                }
-            }
-            ScVal::U64(n) => parts.push(n.to_string()),
-            ScVal::String(s) => parts.push(s.to_utf8_string_lossy()),
-            _ => {}
-        }
-        summaries.push(parts.join(":"));
-    }
-    let s = summaries.join(" | ");
-    if s.len() > 2000 {
-        format!("{}…", &s[..2000])
-    } else {
-        s
-    }
-}
-
-/// Build a per-tx state-changes summary string mirroring what henyey logs from
-/// manager.rs::commit on hash mismatch. Format: "phase:type:size:hash16 | …"
-fn summarize_cdp_meta_changes(meta: Option<&stellar_xdr::curr::TransactionMeta>) -> String {
-    use sha2::{Digest, Sha256};
-    use stellar_xdr::curr::{LedgerEntryChange, LedgerEntryData, TransactionMeta, WriteXdr};
-    let Some(meta) = meta else {
-        return String::new();
-    };
-    let mut parts: Vec<String> = Vec::new();
-    let walk =
-        |changes: &stellar_xdr::curr::LedgerEntryChanges, parts: &mut Vec<String>, phase: &str| {
-            for change in changes.iter() {
-                let entry = match change {
-                    LedgerEntryChange::Created(e) => e,
-                    LedgerEntryChange::Updated(e) => e,
-                    LedgerEntryChange::Restored(e) => e,
-                    _ => continue,
-                };
-                if matches!(entry.data, LedgerEntryData::Ttl(_)) {
-                    continue;
-                }
-                let data_bytes = entry
-                    .data
-                    .to_xdr(stellar_xdr::curr::Limits::none())
-                    .unwrap_or_default();
-                let data_hash = format!("{:x}", Sha256::digest(&data_bytes));
-                parts.push(format!(
-                    "{}:t{:?}:{}:{}",
-                    phase,
-                    std::mem::discriminant(&entry.data),
-                    data_bytes.len(),
-                    &data_hash[..16],
-                ));
-            }
-        };
-    match meta {
-        TransactionMeta::V3(m) => {
-            walk(&m.tx_changes_before, &mut parts, "before");
-            for op in m.operations.iter() {
-                walk(&op.changes, &mut parts, "op");
-            }
-            walk(&m.tx_changes_after, &mut parts, "after");
-        }
-        TransactionMeta::V4(m) => {
-            walk(&m.tx_changes_before, &mut parts, "before");
-            for op in m.operations.iter() {
-                walk(&op.changes, &mut parts, "op");
-            }
-            walk(&m.tx_changes_after, &mut parts, "after");
-        }
-        _ => {}
-    }
-    let mut sorted = parts.clone();
-    sorted.sort();
-    let set_str = sorted.join(" | ");
-    let set_hash = format!("{:x}", Sha256::digest(set_str.as_bytes()));
-    let s = parts.join(" | ");
-    let truncated = if s.len() > 1000 {
-        format!("{}…", &s[..1000])
-    } else {
-        s
-    };
-    format!(
-        "count={} sethash={} {}",
-        parts.len(),
-        &set_hash[..16],
-        truncated
-    )
-}
-
-/// Returns (count, total_bytes) of non-TTL Created/Updated/Restored entries
-/// across a transaction's tx_apply_processing change groups. Mirrors the
-/// stellar-core write_bytes accounting (which excludes TTL entries) for
-/// side-by-side comparison with henyey's `total_write_bytes`.
-fn summarize_cdp_meta(meta: Option<&stellar_xdr::curr::TransactionMeta>) -> (u32, u64) {
-    use stellar_xdr::curr::{LedgerEntryChange, LedgerEntryData, WriteXdr};
-    let Some(meta) = meta else {
-        return (0, 0);
-    };
-    let mut count: u32 = 0;
-    let mut total_bytes: u64 = 0;
-    henyey_common::meta_walk::for_each_change_group(meta, |changes| {
-        for change in changes {
-            let entry = match change {
-                LedgerEntryChange::Created(e) | LedgerEntryChange::Updated(e) => e,
-                LedgerEntryChange::Restored(e) => e,
-                _ => continue,
-            };
-            if matches!(entry.data, LedgerEntryData::Ttl(_)) {
-                continue;
-            }
-            count += 1;
-            if let Ok(bytes) = entry.to_xdr(stellar_xdr::curr::Limits::none()) {
-                total_bytes += bytes.len() as u64;
-            }
-        }
-    });
-    (count, total_bytes)
-}
-
-/// Returns a CSV summary of operation result discriminants for a transaction
-/// result, e.g. "InvokeHostFunction:Trapped". Used to diagnose tx-level
-/// success/fail divergence by exposing per-op outcomes.
-fn describe_op_results(r: &stellar_xdr::curr::TransactionResultResult) -> String {
-    use stellar_xdr::curr::{
-        InnerTransactionResultResult, OperationResult, TransactionResultResult,
-    };
-    fn one_op(op: &OperationResult) -> String {
-        // Print the full Debug repr (without payload values) so that
-        // success/failure variants of inner result types like
-        // InvokeHostFunctionResult::{Success,Trapped,ResourceLimitExceeded,…}
-        // are visible. Truncate to keep log lines reasonable.
-        let s = format!("{:?}", op);
-        if s.len() > 200 {
-            format!("{}…", &s[..200])
-        } else {
-            s
-        }
-    }
-    fn ops_to_csv(ops: &[OperationResult]) -> String {
-        ops.iter().map(one_op).collect::<Vec<_>>().join(",")
-    }
-    match r {
-        TransactionResultResult::TxSuccess(ops) => ops_to_csv(ops.as_slice()),
-        TransactionResultResult::TxFailed(ops) => ops_to_csv(ops.as_slice()),
-        TransactionResultResult::TxFeeBumpInnerSuccess(pair) => {
-            let inner_ops = match &pair.result.result {
-                InnerTransactionResultResult::TxSuccess(ops) => ops_to_csv(ops.as_slice()),
-                InnerTransactionResultResult::TxFailed(ops) => ops_to_csv(ops.as_slice()),
-                other => format!("{:?}", other.discriminant()),
-            };
-            format!("FB+{inner_ops}")
-        }
-        TransactionResultResult::TxFeeBumpInnerFailed(pair) => {
-            let inner_ops = match &pair.result.result {
-                InnerTransactionResultResult::TxSuccess(ops) => ops_to_csv(ops.as_slice()),
-                InnerTransactionResultResult::TxFailed(ops) => ops_to_csv(ops.as_slice()),
-                other => format!("{:?}", other.discriminant()),
-            };
-            format!("FB-{inner_ops}")
-        }
-        other => format!("{:?}", other.discriminant()),
-    }
-}
-
-/// Returns a human-readable name for a `TransactionResultResult` variant.
-fn tx_result_code_name(r: &stellar_xdr::curr::TransactionResultResult) -> String {
-    use stellar_xdr::curr::TransactionResultResult;
-    match r {
-        TransactionResultResult::TxSuccess(_) => "txSuccess".to_string(),
-        TransactionResultResult::TxFailed(_) => "txFailed".to_string(),
-        TransactionResultResult::TxFeeBumpInnerSuccess(_) => "txFeeBumpInnerSuccess".to_string(),
-        TransactionResultResult::TxFeeBumpInnerFailed(_) => "txFeeBumpInnerFailed".to_string(),
-        other => format!("{:?}", other),
-    }
-}
-
-/// Prints pairwise differences between two operation result slices.
-fn print_op_diffs(
-    our_ops: &[stellar_xdr::curr::OperationResult],
-    cdp_ops: &[stellar_xdr::curr::OperationResult],
-) {
-    use stellar_xdr::curr::WriteXdr;
-    for (j, (our_op, cdp_op)) in our_ops.iter().zip(cdp_ops.iter()).enumerate() {
-        let our_op_xdr = our_op
-            .to_xdr(stellar_xdr::curr::Limits::none())
-            .unwrap_or_default();
-        let cdp_op_xdr = cdp_op
-            .to_xdr(stellar_xdr::curr::Limits::none())
-            .unwrap_or_default();
-        if our_op_xdr != cdp_op_xdr {
-            println!("          Op {} differs:", j);
-            println!("            Ours: {:?}", our_op);
-            println!("            CDP:  {:?}", cdp_op);
-        }
-    }
-}
-
-/// Prints all operations from a result slice with a label.
-fn print_all_ops(label: &str, ops: &[stellar_xdr::curr::OperationResult]) {
-    println!("        {} ops ({}):", label, ops.len());
-    for (j, op) in ops.iter().enumerate() {
-        println!("          Op {}: {:?}", j, op);
-    }
-}
-
-/// Prints an exhaustive field-by-field comparison of two ledger headers.
-fn print_header_field_diffs(
-    h: &stellar_xdr::curr::LedgerHeader,
-    c: &stellar_xdr::curr::LedgerHeader,
-    bucket_levels: &[(henyey_common::Hash256, henyey_common::Hash256)],
-) {
-    use henyey_common::Hash256;
-    if h.ledger_version != c.ledger_version {
-        println!(
-            "    DIFF ledger_version: ours={} expected={}",
-            h.ledger_version, c.ledger_version
-        );
-    }
-    if h.previous_ledger_hash != c.previous_ledger_hash {
-        println!(
-            "    DIFF previous_ledger_hash: ours={} expected={}",
-            hex::encode(&h.previous_ledger_hash.0),
-            hex::encode(&c.previous_ledger_hash.0)
-        );
-    }
-    if h.scp_value != c.scp_value {
-        println!("    DIFF scp_value");
-        if h.scp_value.tx_set_hash != c.scp_value.tx_set_hash {
-            println!(
-                "      tx_set_hash: ours={} expected={}",
-                hex::encode(&h.scp_value.tx_set_hash.0),
-                hex::encode(&c.scp_value.tx_set_hash.0)
-            );
-        }
-        if h.scp_value.close_time != c.scp_value.close_time {
-            println!(
-                "      close_time: ours={} expected={}",
-                h.scp_value.close_time.0, c.scp_value.close_time.0
-            );
-        }
-        if h.scp_value.upgrades != c.scp_value.upgrades {
-            println!(
-                "      upgrades: ours={:?} expected={:?}",
-                h.scp_value.upgrades, c.scp_value.upgrades
-            );
-        }
-        if h.scp_value.ext != c.scp_value.ext {
-            println!("      ext: differs");
-        }
-    }
-    let our_bl_hash = Hash256::from(h.bucket_list_hash.0);
-    let expected_bl_hash = Hash256::from(c.bucket_list_hash.0);
-    if our_bl_hash != expected_bl_hash {
-        println!(
-            "    DIFF bucket_list_hash: ours={} expected={}",
-            our_bl_hash.to_hex(),
-            expected_bl_hash.to_hex()
-        );
-        for (i, (curr_hash, snap_hash)) in bucket_levels.iter().enumerate() {
-            println!(
-                "      Level {}: curr={} snap={}",
-                i,
-                curr_hash.to_hex(),
-                snap_hash.to_hex()
-            );
-        }
-    }
-    if h.tx_set_result_hash != c.tx_set_result_hash {
-        println!(
-            "    DIFF tx_set_result_hash: ours={} expected={}",
-            hex::encode(&h.tx_set_result_hash.0),
-            hex::encode(&c.tx_set_result_hash.0)
-        );
-    }
-    if h.ledger_seq != c.ledger_seq {
-        println!(
-            "    DIFF ledger_seq: ours={} expected={}",
-            h.ledger_seq, c.ledger_seq
-        );
-    }
-    if h.total_coins != c.total_coins {
-        println!(
-            "    DIFF total_coins: ours={} expected={}",
-            h.total_coins, c.total_coins
-        );
-    }
-    if h.fee_pool != c.fee_pool {
-        println!(
-            "    DIFF fee_pool: ours={} expected={}",
-            h.fee_pool, c.fee_pool
-        );
-    }
-    if h.inflation_seq != c.inflation_seq {
-        println!(
-            "    DIFF inflation_seq: ours={} expected={}",
-            h.inflation_seq, c.inflation_seq
-        );
-    }
-    if h.id_pool != c.id_pool {
-        println!(
-            "    DIFF id_pool: ours={} expected={}",
-            h.id_pool, c.id_pool
-        );
-    }
-    if h.base_fee != c.base_fee {
-        println!(
-            "    DIFF base_fee: ours={} expected={}",
-            h.base_fee, c.base_fee
-        );
-    }
-    if h.base_reserve != c.base_reserve {
-        println!(
-            "    DIFF base_reserve: ours={} expected={}",
-            h.base_reserve, c.base_reserve
-        );
-    }
-    if h.max_tx_set_size != c.max_tx_set_size {
-        println!(
-            "    DIFF max_tx_set_size: ours={} expected={}",
-            h.max_tx_set_size, c.max_tx_set_size
-        );
-    }
-    if h.skip_list != c.skip_list {
-        println!("    DIFF skip_list:");
-        for (i, (ours, exp)) in h.skip_list.iter().zip(c.skip_list.iter()).enumerate() {
-            if ours != exp {
-                println!(
-                    "      [{}]: ours={} expected={}",
-                    i,
-                    hex::encode(&ours.0),
-                    hex::encode(&exp.0)
-                );
-            }
-        }
-    }
-    if h.ext != c.ext {
-        println!("    DIFF ext: ours={:?} expected={:?}", h.ext, c.ext);
-    }
-}
-
-/// Compare fee bump inner results from both sides and print differences.
-fn print_fee_bump_inner_diffs(
-    our_result: &stellar_xdr::curr::TransactionResultResult,
-    cdp_result: &stellar_xdr::curr::TransactionResultResult,
-) {
-    use stellar_xdr::curr::{InnerTransactionResultResult, TransactionResultResult};
-
-    match (our_result, cdp_result) {
-        (
-            TransactionResultResult::TxFeeBumpInnerFailed(our_inner),
-            TransactionResultResult::TxFeeBumpInnerFailed(cdp_inner),
-        ) => {
-            println!(
-                "        Inner fee: ours={} CDP={}",
-                our_inner.result.fee_charged, cdp_inner.result.fee_charged
-            );
-            let our_inner_code = format!("{:?}", std::mem::discriminant(&our_inner.result.result));
-            let cdp_inner_code = format!("{:?}", std::mem::discriminant(&cdp_inner.result.result));
-            println!(
-                "        Inner result type: ours={} CDP={}",
-                our_inner_code, cdp_inner_code
-            );
-            if let (
-                InnerTransactionResultResult::TxFailed(our_ops),
-                InnerTransactionResultResult::TxFailed(cdp_ops),
-            ) = (&our_inner.result.result, &cdp_inner.result.result)
-            {
-                print_op_diffs(our_ops, cdp_ops);
-                if our_ops.len() != cdp_ops.len() {
-                    println!(
-                        "          Inner op count: ours={} CDP={}",
-                        our_ops.len(),
-                        cdp_ops.len()
-                    );
-                }
-            } else {
-                println!("        Inner result ours: {:?}", our_inner.result.result);
-                println!("        Inner result CDP:  {:?}", cdp_inner.result.result);
-            }
-        }
-        (
-            TransactionResultResult::TxFeeBumpInnerSuccess(our_inner),
-            TransactionResultResult::TxFeeBumpInnerSuccess(cdp_inner),
-        ) => {
-            println!(
-                "        Inner fee: ours={} CDP={}",
-                our_inner.result.fee_charged, cdp_inner.result.fee_charged
-            );
-            if let (
-                InnerTransactionResultResult::TxSuccess(our_ops),
-                InnerTransactionResultResult::TxSuccess(cdp_ops),
-            ) = (&our_inner.result.result, &cdp_inner.result.result)
-            {
-                print_op_diffs(our_ops, cdp_ops);
-            }
-        }
-        (
-            TransactionResultResult::TxFeeBumpInnerSuccess(our_inner),
-            TransactionResultResult::TxFeeBumpInnerFailed(cdp_inner),
-        ) => {
-            println!(
-                "        Inner fee: ours={} CDP={}",
-                our_inner.result.fee_charged, cdp_inner.result.fee_charged
-            );
-            if let InnerTransactionResultResult::TxSuccess(our_ops) = &our_inner.result.result {
-                print_all_ops("Ours inner", our_ops);
-            }
-            if let InnerTransactionResultResult::TxFailed(cdp_ops) = &cdp_inner.result.result {
-                print_all_ops("CDP inner", cdp_ops);
-            } else {
-                println!("        CDP inner result: {:?}", cdp_inner.result.result);
-            }
-        }
-        (
-            TransactionResultResult::TxFeeBumpInnerFailed(our_inner),
-            TransactionResultResult::TxFeeBumpInnerSuccess(cdp_inner),
-        ) => {
-            println!(
-                "        Inner fee: ours={} CDP={}",
-                our_inner.result.fee_charged, cdp_inner.result.fee_charged
-            );
-            println!("        Ours inner result: {:?}", our_inner.result.result);
-            if let InnerTransactionResultResult::TxSuccess(cdp_ops) = &cdp_inner.result.result {
-                print_all_ops("CDP inner", cdp_ops);
-            }
-        }
-        _ => {}
-    }
-}
-
-/// Prints detailed per-TX result diffs between our results and CDP results.
-///
-/// Shows ordering differences, then does a TX-by-TX XDR comparison with
-/// detailed operation-level diffs for all result variant combinations.
-fn print_tx_result_diffs(
-    our_results: &[stellar_xdr::curr::TransactionResultPair],
-    cdp_results: &[stellar_xdr::curr::TransactionResultPair],
-) {
-    use stellar_xdr::curr::{TransactionResultResult, WriteXdr};
-    println!(
-        "    TX count: ours={} CDP={}",
-        our_results.len(),
-        cdp_results.len()
-    );
-
-    // Check if TX ordering differs
-    let mut order_diffs = 0;
-    for (i, (our_tx, cdp_tx)) in our_results.iter().zip(cdp_results.iter()).enumerate() {
-        if our_tx.transaction_hash != cdp_tx.transaction_hash {
-            if order_diffs < 10 {
-                println!(
-                    "    ORDER DIFF at position {}: ours={} CDP={}",
-                    i,
-                    hex::encode(&our_tx.transaction_hash.0),
-                    hex::encode(&cdp_tx.transaction_hash.0)
-                );
-            }
-            order_diffs += 1;
-        }
-    }
-    if order_diffs > 0 {
-        println!("    Total TX ordering differences: {}", order_diffs);
-    } else {
-        println!("    TX ordering is IDENTICAL (same content hashes at every position)");
-    }
-
-    // Detailed TX-by-TX comparison using full XDR
-    let mut diff_count = 0;
-    for (i, (our_tx, cdp_tx)) in our_results.iter().zip(cdp_results.iter()).enumerate() {
-        let our_xdr = our_tx
-            .result
-            .to_xdr(stellar_xdr::curr::Limits::none())
-            .unwrap_or_default();
-        let cdp_xdr = cdp_tx
-            .result
-            .to_xdr(stellar_xdr::curr::Limits::none())
-            .unwrap_or_default();
-
-        if our_xdr != cdp_xdr {
-            diff_count += 1;
-            let our_result = &our_tx.result.result;
-            let cdp_result = &cdp_tx.result.result;
-
-            println!("      TX {}: MISMATCH (XDR differs)", i);
-            println!(
-                "        Result: ours={} CDP={}",
-                tx_result_code_name(our_result),
-                tx_result_code_name(cdp_result)
-            );
-            println!(
-                "        Fee: ours={} CDP={}",
-                our_tx.result.fee_charged, cdp_tx.result.fee_charged
-            );
-            println!(
-                "        TX hash: {}",
-                hex::encode(&our_tx.transaction_hash.0)
-            );
-
-            // Compare operations for same-variant pairs
-            match (our_result, cdp_result) {
-                (
-                    TransactionResultResult::TxFailed(our_ops),
-                    TransactionResultResult::TxFailed(cdp_ops),
-                )
-                | (
-                    TransactionResultResult::TxSuccess(our_ops),
-                    TransactionResultResult::TxSuccess(cdp_ops),
-                ) => {
-                    print_op_diffs(our_ops, cdp_ops);
-                }
-                // One succeeds, other fails — show all ops from both sides
-                (
-                    TransactionResultResult::TxSuccess(our_ops),
-                    TransactionResultResult::TxFailed(cdp_ops),
-                )
-                | (
-                    TransactionResultResult::TxFailed(our_ops),
-                    TransactionResultResult::TxSuccess(cdp_ops),
-                ) => {
-                    print_all_ops("Ours", our_ops);
-                    print_all_ops("CDP", cdp_ops);
-                }
-                _ => {}
-            }
-
-            // Fee bump inner result details
-            print_fee_bump_inner_diffs(our_result, cdp_result);
-
-            // Show CDP ops when ours is TxNotSupported or other non-standard result
-            if !matches!(
-                our_result,
-                TransactionResultResult::TxSuccess(_)
-                    | TransactionResultResult::TxFailed(_)
-                    | TransactionResultResult::TxFeeBumpInnerSuccess(_)
-                    | TransactionResultResult::TxFeeBumpInnerFailed(_)
-            ) {
-                if let TransactionResultResult::TxFailed(cdp_ops) = cdp_result {
-                    println!("        CDP txFailed ops ({}):", cdp_ops.len());
-                    for (j, op) in cdp_ops.iter().enumerate() {
-                        println!("          Op {}: {:?}", j, op);
-                    }
-                }
-            }
-
-            // Limit output to first 10 diffs
-            if diff_count >= 10 {
-                println!("      ... (showing first 10 of potentially more diffs)");
-                break;
-            }
-        }
-    }
-    if diff_count > 0 {
-        println!(
-            "    Total TX diffs: {} out of {}",
-            diff_count,
-            our_results.len().min(cdp_results.len())
-        );
-    }
-}
-
-/// Verifies transaction execution by comparing results against CDP metadata.
-///
-/// Restores bucket list state from a checkpoint, re-executes transactions via
-/// `close_ledger`, and compares results against CDP-produced ledger close metadata.
-pub(crate) async fn cmd_verify_execution(
-    config: AppConfig,
-    opts: VerifyExecutionOptions,
-) -> anyhow::Result<()> {
-    let mut ctx = setup(config, opts).await?;
-    let (mut stats, elapsed) = run_verification_loop(&mut ctx).await?;
-    print_summary(&mut stats, elapsed);
-    if stats.ledgers_mismatched > 0 {
-        anyhow::bail!(
-            "Verification failed with {} mismatched ledgers",
-            stats.ledgers_mismatched
-        );
-    }
-    Ok(())
-}
-
-/// Phase 1-4: Parse config, create clients, download state, initialize LedgerManager.
-async fn setup(config: AppConfig, opts: VerifyExecutionOptions) -> anyhow::Result<VerifyContext> {
-    let VerifyExecutionOptions {
-        from,
-        to,
-        stop_on_error,
-        show_diff,
-        cdp_url,
-        cdp_date,
-        cache_dir,
-        no_cache,
-        quiet,
-    } = opts;
-
-    let init_start = Instant::now();
-
-    if !quiet {
-        println!("Transaction Execution Verification");
-        println!("===================================");
-        println!("Executes transactions via close_ledger and compares against CDP.");
-        println!();
-    }
-
-    // Determine network name
-    let (network_name, is_mainnet) =
-        if config.network.passphrase == "Test SDF Network ; September 2015" {
-            ("testnet", false)
-        } else {
-            ("mainnet", true)
-        };
-
-    // Set network-specific CDP defaults
-    let cdp_url = cdp_url.unwrap_or_else(|| {
-        if is_mainnet {
-            "https://aws-public-blockchain.s3.us-east-2.amazonaws.com/v1.1/stellar/ledgers/pubnet"
-                .to_string()
-        } else {
-            "https://aws-public-blockchain.s3.us-east-2.amazonaws.com/v1.1/stellar/ledgers/testnet"
-                .to_string()
-        }
-    });
-    let cdp_date = cdp_date.unwrap_or_else(|| {
-        if is_mainnet {
-            String::new()
-        } else {
-            "2025-12-18".to_string()
-        }
-    });
-
-    // Determine cache directory
-    let cache_base = if no_cache {
-        None
-    } else {
-        cache_dir.or_else(|| dirs::cache_dir().map(|p| p.join("rs-stellar-core")))
-    };
-
-    // Create archive client
-    let archive = super::first_archive(&config)?;
-
-    if !quiet {
-        println!("Archive: {}", config.history.archives[0].url);
-        let cdp_date_display = if cdp_date.is_empty() {
-            "none (range-based)"
-        } else {
-            &cdp_date
-        };
-        println!("CDP: {} (date: {})", cdp_url, cdp_date_display);
-        if let Some(ref cache) = cache_base {
-            println!("Cache: {}", cache.display());
-        } else {
-            println!("Cache: disabled");
-        }
-    }
-
-    // Get current ledger and calculate range
-    let root_has = archive.fetch_root_has().await?;
-    let current_ledger = root_has.current_ledger;
-
-    let end_ledger = to.unwrap_or_else(|| {
-        checkpoint::latest_checkpoint_before_or_at(current_ledger).unwrap_or(current_ledger)
-    });
-    let start_ledger = from.unwrap_or_else(|| {
-        let freq = henyey_history::checkpoint_frequency();
-        checkpoint::checkpoint_containing(end_ledger)
-            .saturating_sub(4 * freq)
-            .max(freq)
-    });
-
-    let freq = henyey_history::checkpoint_frequency();
-    let init_checkpoint =
-        checkpoint::latest_checkpoint_before_or_at(start_ledger.saturating_sub(1))
-            .unwrap_or(freq - 1);
-    let end_checkpoint = checkpoint::checkpoint_containing(end_ledger);
-
-    if !quiet {
-        println!("Ledger range: {} to {}", start_ledger, end_ledger);
-        println!("Initial state: checkpoint {}", init_checkpoint);
-        println!();
-    }
-
-    // Create CDP client with caching
-    let (_cdp_dir_holder, cdp) = if let Some(ref cache) = cache_base {
-        let cdp = CachedCdpDataLake::new(&cdp_url, &cdp_date, cache, network_name)?;
-        (None, cdp)
-    } else {
-        let temp = tempfile::tempdir()?;
-        let cdp = CachedCdpDataLake::new(&cdp_url, &cdp_date, temp.path(), network_name)?;
-        (Some(temp), cdp)
-    };
-
-    // Setup bucket manager
-    let (_bucket_dir_holder, bucket_path) = if let Some(ref cache) = cache_base {
-        let path = cache.join("buckets").join(network_name);
-        std::fs::create_dir_all(&path)?;
-        (None, path)
-    } else {
-        let temp = tempfile::tempdir()?;
-        let path = temp.path().to_path_buf();
-        (Some(temp), path)
-    };
-    let bucket_manager = Arc::new(BucketManager::with_persist_index(
-        bucket_path.clone(),
-        true,
-    )?);
-
-    // Download initial state
-    if !quiet {
-        println!(
-            "Downloading initial state at checkpoint {}...",
-            init_checkpoint
-        );
-    }
-    let init_has = archive.fetch_checkpoint_has(init_checkpoint).await?;
-
-    // Extract bucket hashes
-    let bucket_hashes = init_has.bucket_hash_pairs();
-
-    let live_next_states: Vec<Option<PendingMergeState>> = init_has.live_next_states()?;
-
-    // Extract hot archive bucket hashes (protocol 23+)
-    let hot_archive_hashes = init_has.hot_archive_bucket_hash_pairs();
-
-    let hot_archive_next_states: Option<Vec<Option<PendingMergeState>>> =
-        init_has.hot_archive_next_states()?;
-
-    // Collect all bucket hashes to download
-    let mut all_hashes: Vec<Hash256> = Vec::new();
-    for (curr, snap) in &bucket_hashes {
-        all_hashes.push(*curr);
-        all_hashes.push(*snap);
-    }
-    for merge_state in live_next_states.iter().flatten() {
-        all_hashes.extend(merge_state.referenced_hashes().copied());
-    }
-    if let Some(ref ha_hashes) = hot_archive_hashes {
-        for (curr, snap) in ha_hashes {
-            all_hashes.push(*curr);
-            all_hashes.push(*snap);
-        }
-    }
-    if let Some(ref ha_states) = hot_archive_next_states {
-        for merge_state in ha_states.iter().flatten() {
-            all_hashes.extend(merge_state.referenced_hashes().copied());
-        }
-    }
-    let all_hashes: Vec<&Hash256> = all_hashes.iter().filter(|h| !h.is_zero()).collect();
-
-    // Download buckets
-    let (cached, downloaded) =
-        super::download_buckets_parallel(&archive, bucket_manager.clone(), all_hashes).await?;
-    println!(
-        "[INIT] Bucket download: {} cached, {} downloaded",
-        cached, downloaded
-    );
-
-    // Restore bucket lists
-    let mut bucket_list =
-        BucketList::restore_from_has(&bucket_hashes, &live_next_states, |hash| {
-            bucket_manager.load_bucket(hash).map(|b| (*b).clone())
-        })?;
-    bucket_list.set_bucket_dir(bucket_manager.bucket_dir().to_path_buf());
-
-    let mut hot_archive_bucket_list = match (&hot_archive_hashes, &hot_archive_next_states) {
-        (Some(ref hashes), Some(ref next_states)) => {
-            HotArchiveBucketList::restore_from_has(hashes, next_states, |hash| {
-                bucket_manager.load_hot_archive_bucket(hash)
-            })?
-        }
-        _ => HotArchiveBucketList::new(),
-    };
-
-    // Get init header and restart merges
-    let init_headers = archive.fetch_ledger_headers(init_checkpoint).await?;
-    let init_header_entry = init_headers
-        .iter()
-        .find(|h| h.header.ledger_seq == init_checkpoint);
-    let init_protocol_version = init_header_entry
-        .map(|h| h.header.ledger_version)
-        .unwrap_or(25);
-
-    // Enable structure-based merge restarts to match stellar-core online mode behavior.
-    //
-    // Although stellar-core standalone offline commands skip restartMerges, we are comparing
-    // our results against CDP headers produced by stellar-core in ONLINE mode. The stellar-core node
-    // that produced those headers had full structure-based merge restarts enabled.
-    //
-    // In stellar-core online mode, restartMerges uses mLevels[i-1].getSnap() (the old snap
-    // from HAS) to start merges. Without structure-based restarts, add_batch would
-    // use snap() which returns the snapped curr (different input!).
-    bucket_list
-        .restart_merges_from_has(
-            init_checkpoint,
-            init_protocol_version,
-            &live_next_states,
-            |hash| bucket_manager.load_bucket_for_merge(hash),
-            true, // restart_structure_based = true to match stellar-core online mode
-        )
-        .await?;
-
-    if let Some(ref ha_next_states) = hot_archive_next_states {
-        hot_archive_bucket_list.restart_merges_from_has(
-            init_checkpoint,
-            init_protocol_version,
-            ha_next_states,
-            |hash| bucket_manager.load_hot_archive_bucket_for_merge(hash),
-            true, // restart_structure_based = true to match stellar-core online mode
-        )?;
-    }
-
-    // Create and initialize LedgerManager
-    let mut ledger_manager = LedgerManager::new(
-        config.network.passphrase.clone(),
-        LedgerManagerConfig {
-            validate_bucket_hash: true,
-            bucket_list_db: config.buckets.bucket_list_db.clone(),
-            // Use single-threaded scan to reduce peak memory during initialization.
-            // Verify-execution runs on memory-constrained CI runners where concurrent
-            // level scans would cause OOM.
-            scan_thread_count: 1,
-            ..Default::default()
-        },
-    );
-
-    // Wire merge map for bucket merge deduplication during replay.
-    let finished_merges =
-        std::sync::Arc::new(std::sync::RwLock::new(henyey_bucket::BucketMergeMap::new()));
-    ledger_manager.set_merge_map(finished_merges);
-
-    let init_header_entry = init_header_entry
-        .ok_or_else(|| anyhow::anyhow!("No header found for checkpoint {}", init_checkpoint))?;
-    let init_header_hash = verify::verify_ledger_header_history_entry(&init_header_entry)?;
-    ledger_manager.initialize(
-        bucket_list,
-        hot_archive_bucket_list,
-        init_header_entry.header.clone(),
-        init_header_hash,
-    )?;
-
-    println!(
-        "[INIT] TOTAL initialization: {:.2}s",
-        init_start.elapsed().as_secs_f64()
-    );
-
-    Ok(VerifyContext {
-        archive,
-        cdp,
-        ledger_manager,
-        _bucket_manager: bucket_manager,
-        _cdp_dir_holder,
-        _bucket_dir_holder,
-        start_ledger,
-        end_ledger,
-        init_checkpoint,
-        end_checkpoint,
-        init_header_hash,
-        stop_on_error,
-        show_diff,
-        quiet,
-    })
-}
+use super::diffs::*;
+use super::{VerifyContext, VerifyStats};
 
 /// Phase 5: Main verification loop — iterate checkpoints and verify each ledger.
-async fn run_verification_loop(ctx: &mut VerifyContext) -> anyhow::Result<(VerifyStats, Duration)> {
+pub(super) async fn run_verification_loop(
+    ctx: &mut VerifyContext,
+) -> anyhow::Result<(VerifyStats, Duration)> {
     let mut stats = VerifyStats::default();
     let mut prev_ledger_hash = ctx.init_header_hash;
 
@@ -1491,7 +431,7 @@ async fn verify_single_ledger(
 }
 
 /// Phase 6: Print verification and performance summaries.
-fn print_summary(stats: &mut VerifyStats, elapsed: Duration) {
+pub(super) fn print_summary(stats: &mut VerifyStats, elapsed: Duration) {
     // Print summary
     println!();
     println!();
